+++ gradle.properties
@@ -14,6 +14,6 @@ kotlin.mpp.enableGranularSourceSetsMetadata=true
 kotlin.native.binary.memoryModel=experimental
 kotlin.version=1.8.0
 agp.version=7.1.3
-compose.version=1.3.0
+compose.version=1.5.0-dev1049
 ktor.version=2.2.1
 kotlin.js.ir.output.granularity=whole-program
