+++ shared/src/androidMain/kotlin/com/example/travelapp_kmp/ImageBitmap.common.android.kt
@@ -1,0 +1,12 @@
+package com.example.travelapp_kmp
+
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.asImageBitmap
+
+actual fun ByteArray.toImageBitmap(): ImageBitmap = toAndroidBitmap().asImageBitmap()
+
+fun ByteArray.toAndroidBitmap(): Bitmap {
+    return BitmapFactory.decodeByteArray(this, 0, size)
+}
