--- shared/src/commonMain/kotlin/com/example/traveapp_kmp/TravelAppTheme.kt
@@ -1,8 +1,0 @@
-package com.example.traveapp_kmp
-
-import androidx.compose.runtime.Composable
-
-@Composable
-internal fun TravelAppTheme(content: @Composable () -> Unit) {
-    content()
-}
