--- shared/src/commonMain/kotlin/com/example/traveapp_kmp/network/CountriesApi.kt
@@ -1,7 +1,0 @@
-package com.example.traveapp_kmp.network
-
-import com.example.traveapp_kmp.listing.Country
-
-interface CountriesApi {
-    suspend fun getCountriesList(): List<Country>
-}
