+++ shared/src/commonMain/kotlin/com/example/travelapp_kmp/ImageBitmap.common.kt
@@ -1,0 +1,5 @@
+package com.example.travelapp_kmp
+
+import androidx.compose.ui.graphics.ImageBitmap
+
+expect fun ByteArray.toImageBitmap(): ImageBitmap
