+++ shared/src/commonMain/kotlin/com/example/travelapp_kmp/TravelAppTheme.kt
@@ -1,0 +1,8 @@
+package com.example.travelapp_kmp
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun TravelAppTheme(content: @Composable () -> Unit) {
+    content()
+}
