+++ shared/src/commonMain/kotlin/com/example/travelapp_kmp/network/CountriesApi.kt
@@ -1,0 +1,7 @@
+package com.example.travelapp_kmp.network
+
+import com.example.travelapp_kmp.listing.Country
+
+interface CountriesApi {
+    suspend fun getCountriesList(): List<Country>
+}
