+++ shared/src/iosMain/kotlin/com/example/travelapp_kmp/AppViewiOS.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.toSize
-import com.example.traveapp_kmp.CommonView
+import com.example.travelapp_kmp.CommonView
 import com.seiko.imageloader.ImageLoader
 import com.seiko.imageloader.LocalImageLoader
 import com.seiko.imageloader.cache.memory.maxSizePercent
