+++ shared/src/iosMain/kotlin/com/example/travelapp_kmp/ImageBitmap.common.ios.kt
@@ -1,0 +1,8 @@
+package com.example.travelapp_kmp
+
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.toComposeImageBitmap
+import org.jetbrains.skia.Image
+
+actual fun ByteArray.toImageBitmap(): ImageBitmap =
+    Image.makeFromEncoded(this).toComposeImageBitmap()
