+++ shared/src/macosMain/kotlin/com/example/travelapp_kmp/ImageBitmap.common.macos.kt
@@ -1,0 +1,9 @@
+package com.example.travelapp_kmp
+
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.toComposeImageBitmap
+import org.jetbrains.skia.Image
+
+actual fun ByteArray.toImageBitmap(): ImageBitmap {
+    return Image.makeFromEncoded(this).toComposeImageBitmap()
+}
